[ISSUE #24] park interface complete. (#26)
* seata and ss transaction complete

* 场库列表接口完成

+++ shared-trad-customer-serv/shared-trad-customer-impl/src/main/java/cn/suparking/customer/model/park/Location.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Location {
 
-    private Long latitude;
+    private Double latitude;
 
-    private Long longitude;
+    private Double longitude;
 }
